+++ tesseraql-maven-plugin/src/main/java/io/tesseraql/maven/TestMojo.java
@@ -1,6 +1,8 @@
 package io.tesseraql.maven;
 
 import io.tesseraql.identity.RealmConfig;
+import io.tesseraql.report.AppTestRunner;
+import io.tesseraql.report.DriverManagerDataSource;
 import io.tesseraql.test.TestReport;
 import java.io.File;
 import org.apache.maven.plugin.AbstractMojo;